Remove bindToContext API
