+++ packages/framework/aqueduct/package.json
@@ -111,6 +111,9 @@
 			},
 			"ClassDeclaration_PureDataObject": {
 				"backCompat": false
+			},
+			"InterfaceDeclaration_IDataObjectProps": {
+				"backCompat": false
 			}
 		}
 	}
