+++ packages/runtime/container-runtime-definitions/package.json
@@ -53,6 +53,11 @@
 		"typescript": "~4.5.5"
 	},
 	"typeValidation": {
-		"broken": {}
+		"broken": {
+			"RemovedInterfaceDeclaration_IDataStoreWithBindToContext_Deprecated": {
+				"forwardCompat": false,
+				"backCompat": false
+			}
+		}
 	}
 }
