+++ packages/runtime/container-runtime-definitions/src/index.ts
@@ -7,5 +7,4 @@ export {
 	IContainerRuntime,
 	IContainerRuntimeBaseWithCombinedEvents,
 	IContainerRuntimeEvents,
-	IDataStoreWithBindToContext_Deprecated,
 } from "./containerRuntime";
