+++ packages/runtime/container-runtime-definitions/src/test/types/validateContainerRuntimeDefinitionsPrevious.generated.ts
@@ -88,23 +88,11 @@ use_old_InterfaceDeclaration_IContainerRuntimeEvents(
 /*
 * Validate forward compat by using old type in place of current type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "InterfaceDeclaration_IDataStoreWithBindToContext_Deprecated": {"forwardCompat": false}
+* "RemovedInterfaceDeclaration_IDataStoreWithBindToContext_Deprecated": {"forwardCompat": false}
 */
-declare function get_old_InterfaceDeclaration_IDataStoreWithBindToContext_Deprecated():
-    TypeOnly<old.IDataStoreWithBindToContext_Deprecated>;
-declare function use_current_InterfaceDeclaration_IDataStoreWithBindToContext_Deprecated(
-    use: TypeOnly<current.IDataStoreWithBindToContext_Deprecated>);
-use_current_InterfaceDeclaration_IDataStoreWithBindToContext_Deprecated(
-    get_old_InterfaceDeclaration_IDataStoreWithBindToContext_Deprecated());
 
 /*
 * Validate back compat by using current type in place of old type
 * If breaking change required, add in package.json under typeValidation.broken:
-* "InterfaceDeclaration_IDataStoreWithBindToContext_Deprecated": {"backCompat": false}
+* "RemovedInterfaceDeclaration_IDataStoreWithBindToContext_Deprecated": {"backCompat": false}
 */
-declare function get_current_InterfaceDeclaration_IDataStoreWithBindToContext_Deprecated():
-    TypeOnly<current.IDataStoreWithBindToContext_Deprecated>;
-declare function use_old_InterfaceDeclaration_IDataStoreWithBindToContext_Deprecated(
-    use: TypeOnly<old.IDataStoreWithBindToContext_Deprecated>);
-use_old_InterfaceDeclaration_IDataStoreWithBindToContext_Deprecated(
-    get_current_InterfaceDeclaration_IDataStoreWithBindToContext_Deprecated());
