+++ packages/runtime/test-runtime-utils/src/mocksDataStoreContext.ts
@@ -109,10 +109,6 @@ export class MockFluidDataStoreContext implements IFluidDataStoreContext {
 		throw new Error("Method not implemented.");
 	}
 
-	public bindToContext(): void {
-		throw new Error("Method not implemented.");
-	}
-
 	public setChannelDirty(address: string): void {
 		throw new Error("Method not implemented.");
 	}
