+++ packages/test/test-end-to-end-tests/src/test/rootDatastores.spec.ts
@@ -12,10 +12,7 @@ import {
 	SummaryCollection,
 	DefaultSummaryConfiguration,
 } from "@fluidframework/container-runtime";
-import {
-	IContainerRuntime,
-	IDataStoreWithBindToContext_Deprecated,
-} from "@fluidframework/container-runtime-definitions";
+import { IContainerRuntime } from "@fluidframework/container-runtime-definitions";
 import { FluidObject, IFluidRouter } from "@fluidframework/core-interfaces";
 import { requestFluidObject } from "@fluidframework/runtime-utils";
 import {
@@ -393,61 +390,5 @@ describeFullCompat("Named root data stores", (getTestObjectProvider) => {
 				assert.ok(await getAliasedDataStoreEntryPoint(dataObject3, alias));
 			},
 		);
-
-		/**
-		 * Aliasing datastores summarized before the alias op is sent and after the attach op is sent
-		 * does not cause a datastore corruption issue
-		 *
-		 * This test validates a bug where the rootiness of a datastore was not set to true in the
-		 * above scenario.
-		 */
-		it("Aliasing a bound datastore marks it as root correctly", async () => {
-			const aliasableDataStore1 = await runtimeOf(dataObject1).createDataStore(packageName);
-			const aliasedDataStoreResponse1 = await aliasableDataStore1.request({ url: "/" });
-			const aliasedDataStore1 = aliasedDataStoreResponse1.value as ITestFluidObject;
-			// Casting any to repro a race condition where bindToContext is called before summarization,
-			// but aliasing happens afterwards
-			(
-				aliasableDataStore1 as IDataStoreWithBindToContext_Deprecated
-			).fluidDataStoreChannel?.bindToContext?.();
-			await provider.ensureSynchronized();
-
-			let dataStore: FluidObject | undefined;
-			let error: unknown;
-			try {
-				dataStore = await getAliasedDataStoreEntryPoint(
-					dataObject2,
-					aliasedDataStore1.runtime.id,
-				);
-			} catch (e) {
-				// back-compat - getRootDataStore throws an error if the data store doesn't exist.
-				error = e;
-			}
-			assert(
-				dataStore === undefined || error !== undefined,
-				"Expected getAliasedDataStoreEntryPoint to fail as the datastore is not yet a root datastore",
-			);
-
-			// Alias a datastore
-			const _alias = "alias";
-			const aliasResult1 = await aliasableDataStore1.trySetAlias(_alias);
-			assert(
-				aliasResult1 === "Success",
-				`Expected an successful aliasing. Got: ${aliasResult1}`,
-			);
-			await provider.ensureSynchronized();
-
-			// Should be able to retrieve root datastore from remote
-			assert.doesNotThrow(
-				async () => getAliasedDataStoreEntryPoint(dataObject2, _alias),
-				"A remote aliased datastore should be a root datastore",
-			);
-
-			// Should be able to retrieve local root datastore
-			assert.doesNotThrow(
-				async () => getAliasedDataStoreEntryPoint(dataObject1, _alias),
-				"A local aliased datastore should be a root datastore",
-			);
-		});
 	});
 });
